+++ sdk/schemaregistry/azure-schemaregistry/azure/schemaregistry/_common/_response_handlers.py
@@ -34,7 +34,7 @@ def _parse_schema_properties_dict(response):
 def _parse_response_schema_properties(response, format):
     # pylint:disable=redefined-builtin
     properties_dict = _parse_schema_properties_dict(response)
-    properties_dict["format"] = SchemaFormat(format.lower())
+    properties_dict["format"] = SchemaFormat(format)
     return SchemaProperties(**properties_dict)
 
 
@@ -42,7 +42,7 @@ def _parse_response_schema(response):
     # pylint:disable=redefined-builtin
     schema_props_dict = _parse_schema_properties_dict(response)
     format = response.headers.get("content-type").split("serialization=")[1]
-    schema_props_dict["format"] = SchemaFormat(format.lower())
+    schema_props_dict["format"] = SchemaFormat(format)
     return Schema(
         definition=response.text(), properties=SchemaProperties(**schema_props_dict)
     )