+++ sdk/schemaregistry/azure-schemaregistry/CHANGELOG.md
@@ -13,7 +13,7 @@
 - `SchemaProperties` has the following instance variables: `id` and `format`:
   - The type of `format` has been changed from `str` to `SchemaFormat`.
 - `Schema` has the following properties: `properties` and `definition`.
-- `SchemaFormat` provides the schema format to be stored by the service. Currently, the only supported format is `avro`.
+- `SchemaFormat` provides the schema format to be stored by the service. Currently, the only supported format is `Avro`.
 - `api_version` has been added as a keyword arg to the sync and async `SchemaRegistryClient` constructors.
 
 ### Breaking Changes